@@ -1,8 +1,8 @@
 import { Route, Routes } from 'react-router-dom';
 import { Login } from 'components';
-import Home from 'containers/Home';
+import { Home } from 'containers/home';
 
-const App = (): JSX.Element => {
+const App = () => {
   return (
     <Routes>
       <Route path='/login' element={<Login />} />
@@ -11,4 +11,4 @@ const App = (): JSX.Element => {
   );
 };
 
-export default App;
+export { App };
