@@ -1,16 +1,16 @@
-import SanityClientConstructor from '@sanity/client';
 import imageUrlBuilder from '@sanity/image-url/';
 import { SanityImageSource } from '@sanity/image-url/lib/types/types';
 import { v4 as uuidv4 } from 'uuid';
+import { createClient } from '@sanity/client';
 
-export const client = SanityClientConstructor({
-  projectId: process.env.REACT_APP_SANITY_PROJECT_ID,
+export const client = createClient({
+  projectId: import.meta.env.VITE_REACT_APP_SANITY_PROJECT_ID,
   dataset: 'production',
   apiVersion: '2021-11-16',
   // set useCdn to `false` to prevent stale images when content is updated
   useCdn: false,
   // for a real app never EVER put a sensitive access token in client-side code!
-  token: process.env.REACT_APP_SANITY_TOKEN,
+  token: import.meta.env.VITE_REACT_APP_SANITY_TOKEN,
   // we need to turn the warning off because this isn't a real app so we're exposing the token
   ignoreBrowserTokenWarning: true,
 });
